FEAT and FIX

+++ app/Http/Controllers/BackgroundController.php
@@ -6,6 +6,7 @@ use App\Story;
 use App\Scene;
 use App\Background;
 use Illuminate\Http\Request;
+use App\Providers\HelperServiceProvider as Helpers;
 
 class BackgroundController extends Controller
 {
@@ -22,6 +23,10 @@ class BackgroundController extends Controller
 	public function index($story_id)
     {		
 		$story = Story::find($story_id);
+		if (Helpers::checkPermission($story_id) == false){
+			return view('errors/403',  array());
+			exit();		
+		}
 		$backgrounds = $story->backgrounds();
         return view('background/index', compact("backgrounds","story"));
     }
@@ -30,6 +35,10 @@ class BackgroundController extends Controller
 	{	
 		$background = new Background();
 		$story = Story::find($story_id);
+		if (Helpers::checkPermission($story_id) == false){
+			return view('errors/403',  array());
+			exit();		
+		}
 		$method = "POST";
 		return view('background/edit',compact('background','method','story'));
 	}
@@ -44,6 +53,10 @@ class BackgroundController extends Controller
 	public function show($id)
 	{
 		$background = Background::find($id);
+		if (Helpers::checkPermission($background->story_id) == false){
+			return view('errors/403',  array());
+			exit();		
+		}
 		return view('background/show',compact('background'));
 	}
 
@@ -54,15 +67,34 @@ class BackgroundController extends Controller
 			$background->name = $inputs["name"];			
 		}
 		
+		$background->picture = "";
 		if (isset($inputs["picture"])){
 			$background->picture = $inputs["picture"];			
 		}
-		
-		//TODO: check permission
+				
 		if (isset($inputs["story_id"])){
 			$background->story_id = $inputs["story_id"];
 		}
 
+		if (Helpers::checkPermission($background->story_id) == false){
+			return view('errors/403',  array());
+			exit();		
+		}
+		
+		//Upload file
+		if ($request->file("picture_file") != ""){
+			if (substr(strtolower($request->file("picture_file")->getClientOriginalName()),-4) == ".png"){
+				if (!is_dir("stories")){
+					mkdir ("stories");
+				}
+				if (!is_dir("stories/".$background->story_id)){
+					mkdir ("stories/".$background->story_id);
+				}
+				
+				Storage::disk('public')->put("stories/".$background->story_id."/".Helpers::encName($background->name).".png", file_get_contents($request->file("picture_file")));			
+			}
+		}
+		
 		$background->save();
 		
 		return $background;
@@ -71,6 +103,10 @@ class BackgroundController extends Controller
 	public function edit(Request $request, $id)
 	{	
 		$background = Background::find($id);
+		if (Helpers::checkPermission($background->story_id) == false){
+			return view('errors/403',  array());
+			exit();		
+		}
 		$story = Story::find($background->story_id);
 		$method = "PUT";
 		return view('background/edit',compact('background','story','method'));
@@ -79,12 +115,21 @@ class BackgroundController extends Controller
 	public function update(Request $request, $id)
 	{		
 		$background = Background::find($id);
+		if (Helpers::checkPermission($background->story_id) == false){
+			return view('errors/403',  array());
+			exit();		
+		}
 		$background = $this->save($background, $request);
 		return redirect('story/'.$background->story_id.'/background')->withOk("The background " . $background->name . " has been saved .");
 	}
 	
 	public function destroy($background_id)
 	{	
+		$background = Background::find($background_id);
+		if (Helpers::checkPermission($background->story_id) == false){
+			return view('errors/403',  array());
+			exit();		
+		}
 		Background::destroy($background_id);
 		return redirect()->back();
 	}
